+++ docs/2-resource-customization.md
@@ -10,13 +10,24 @@ resource you must first create a Rails model for it.
 ## Create a Resource
 
 The basic command for creating a resource is `rails g active_admin:resource Post`.
-The generator will produce an empty `app/admin/posts.rb` file like so:
+The generator will produce a `app/admin/posts.rb` file like the following:
 
 ```ruby
 ActiveAdmin.register Post do
-  # everything happens here :D
+  permit_params :title
+
+  filter :title
+  filter :created_at
+  filter :updated_at
+
+  actions :all, except: []
+
+  # index, show, form ...
 end
 ```
+
+The generator will try to determine possible fields for each section as best
+as possible but you may need to tweak further to get started.
 
 ## Setting up Strong Parameters
