+++ lib/generators/active_admin/page/page_generator.rb
@@ -7,7 +7,6 @@ module ActiveAdmin
       def generate_config_file
         template "page.rb", "app/admin/#{file_path.tr('/', '_')}.rb"
       end
-
     end
   end
 end
