+++ lib/generators/active_admin/resource/resource_generator.rb
@@ -1,21 +1,60 @@
 # frozen_string_literal: true
-require "active_admin/generators/boilerplate"
-
 module ActiveAdmin
   module Generators
     class ResourceGenerator < Rails::Generators::NamedBase
       desc "Registers resources with Active Admin"
 
-      class_option :include_boilerplate, type: :boolean, default: false,
-                                         desc: "Generate boilerplate code for your resource."
-
       source_root File.expand_path("templates", __dir__)
 
       def generate_config_file
-        @boilerplate = ActiveAdmin::Generators::Boilerplate.new(class_name)
-        template "admin.rb.erb", "app/admin/#{file_path.tr('/', '_').pluralize}.rb"
+        template "resource.rb.erb", "app/admin/#{file_path.tr('/', '_').pluralize}.rb"
       end
 
+      protected
+
+      def attributes
+        @attributes ||= class_name.constantize.new.attributes.keys
+      end
+
+      def assignable_attributes
+        @assignable_attributes ||= attributes - %w(id created_at updated_at)
+      end
+
+      def permit_params
+        assignable_attributes.map { |a| a.to_sym.inspect }.join(", ")
+      end
+
+      def rows
+        attributes.map { |a| row(a) }.join("\n      ")
+      end
+
+      def row(name)
+        "row :#{name.gsub(/_id$/, '')}"
+      end
+
+      def columns
+        attributes.map { |a| column(a) }.join("\n    ")
+      end
+
+      def column(name)
+        "column :#{name.gsub(/_id$/, '')}"
+      end
+
+      def filters
+        attributes.map { |a| filter(a) }.join("\n  ")
+      end
+
+      def filter(name)
+        "filter :#{name.gsub(/_id$/, '')}"
+      end
+
+      def form_inputs
+        assignable_attributes.map { |a| form_input(a) }.join("\n      ")
+      end
+
+      def form_input(name)
+        "f.input :#{name.gsub(/_id$/, '')}"
+      end
     end
   end
 end
