--- lib/generators/active_admin/resource/templates/admin.rb.erb
@@ -1,46 +1,0 @@
-ActiveAdmin.register <%= class_name %> do
-
-  # See permitted parameters documentation:
-  # https://github.com/activeadmin/activeadmin/blob/master/docs/2-resource-customization.md#setting-up-strong-parameters
-  #
-  # Uncomment all parameters which should be permitted for assignment
-  #
-  # permit_params <%= @boilerplate.permit_params %>
-  #
-  # or
-  #
-  # permit_params do
-  #   permitted = [<%= @boilerplate.permit_params %>]
-  #   permitted << :other if params[:action] == 'create' && current_user.admin?
-  #   permitted
-  # end
-  <% if options.include_boilerplate? %>
-  # Limit actions available to your users by adding them to the 'except' array
-  # actions :all, except: []
-
-  # Add or remove filters (you can use any ActiveRecord scope) to toggle their
-  # visibility in the sidebar
-  <%= @boilerplate.filters %>
-
-  # Add or remove columns to toggle their visiblity in the index action
-  # index do
-  #   selectable_column
-  #   id_column
-  <%= @boilerplate.columns %>
-  #   actions
-  # end
-
-  # Add or remove rows to toggle their visiblity in the show action
-  # show do |<%= class_name.downcase %>|
-  <%= @boilerplate.rows %>
-  # end
-
-  # Add or remove fields to toggle their visibility in the form
-  # form do |f|
-  #   f.inputs do
-  <%= @boilerplate.form_inputs %>
-  #   end
-  #   f.actions
-  # end
-  <% end %>
-end
